+++ src/main/java/sayTheSpire/ui/effects/handlers/CardManipulationEffectHandler.java
@@ -7,18 +7,17 @@ import com.megacrit.cardcrawl.vfx.FastCardObtainEffect;
 import basemod.ReflectionHacks;
 import sayTheSpire.ui.effects.EffectHandler;
 import sayTheSpire.ui.effects.EffectManager;
+import sayTheSpire.events.CombatCardTextEvent;
+import sayTheSpire.events.Event;
+import sayTheSpire.events.ObtainTextEvent;
 import sayTheSpire.events.TextEvent;
+import sayTheSpire.utils.OutputUtils;
 import sayTheSpire.Output;
 
 public class CardManipulationEffectHandler extends EffectHandler {
 
-    private enum ManipulationType {
-        EXHAUST, PURGE, SHOW_ADD_DISCARD, SHOW_ADD_DRAW, SHOW_ADD_HAND, SHOW_CARD_OBTAIN, SHOW_CARD_BRIEFLY,
-        FAST_OBTAIN, INVALID
-    };
-
-    private ManipulationType manipulationType;
     private AbstractCard card;
+    private Event event;
 
     public CardManipulationEffectHandler(AbstractGameEffect effect) {
         super(effect);
@@ -27,62 +26,40 @@ public class CardManipulationEffectHandler extends EffectHandler {
 
     private void setup() {
         if (this.effect instanceof ExhaustCardEffect) {
-            this.manipulationType = ManipulationType.EXHAUST;
             this.card = (AbstractCard) ReflectionHacks.getPrivate(this.effect, ExhaustCardEffect.class, "c");
+            this.event = new CombatCardTextEvent(this.card.name + " exhausted");
         } else if (this.effect instanceof PurgeCardEffect) {
-            this.manipulationType = ManipulationType.PURGE;
             this.card = (AbstractCard) ReflectionHacks.getPrivate(this.effect, PurgeCardEffect.class, "card");
+            this.event = new TextEvent(this.card.name + " removed from your deck");
         } else if (this.effect instanceof ShowCardAndAddToDiscardEffect) {
-            this.manipulationType = ManipulationType.SHOW_ADD_DISCARD;
             this.card = (AbstractCard) ReflectionHacks.getPrivate(this.effect, ShowCardAndAddToDiscardEffect.class,
                     "card");
+            this.event = new CombatCardTextEvent(this.card.name + " added to discard");
         } else if (this.effect instanceof ShowCardAndAddToDrawPileEffect) {
-            this.manipulationType = ManipulationType.SHOW_ADD_DRAW;
             this.card = (AbstractCard) ReflectionHacks.getPrivate(this.effect, ShowCardAndAddToDrawPileEffect.class,
                     "card");
+            this.event = new CombatCardTextEvent(this.card.name + " added to draw");
         } else if (this.effect instanceof ShowCardAndAddToHandEffect) {
-            this.manipulationType = ManipulationType.SHOW_ADD_HAND;
             this.card = (AbstractCard) ReflectionHacks.getPrivate(this.effect, ShowCardAndAddToHandEffect.class,
                     "card");
+            this.event = new CombatCardTextEvent(this.card.name + " added to hand");
         } else if (this.effect instanceof ShowCardAndObtainEffect) {
-            this.manipulationType = ManipulationType.SHOW_CARD_OBTAIN;
             this.card = (AbstractCard) ReflectionHacks.getPrivate(this.effect, ShowCardAndObtainEffect.class, "card");
+            this.event = new ObtainTextEvent(this.card.name + " added to deck");
         } else if (this.effect instanceof ShowCardBrieflyEffect) {
-            this.manipulationType = ManipulationType.SHOW_CARD_BRIEFLY;
             this.card = (AbstractCard) ReflectionHacks.getPrivate(this.effect, ShowCardBrieflyEffect.class, "card");
+            this.event = new TextEvent(this.card.name + " shown briefly");
         } else if (this.effect instanceof FastCardObtainEffect) {
-            this.manipulationType = ManipulationType.FAST_OBTAIN;
             this.card = (AbstractCard) ReflectionHacks.getPrivate(this.effect, FastCardObtainEffect.class, "card");
+            this.event = new ObtainTextEvent(this.card.name + " added to deck");
         } else {
-            this.manipulationType = ManipulationType.INVALID;
             this.card = null;
-        }
-    }
-
-    public String getMessage() {
-        switch (this.manipulationType) {
-        case EXHAUST:
-            return this.card.name + " exhausted";
-        case PURGE:
-            return this.card.name + " removed from your deck";
-        case SHOW_ADD_DISCARD:
-            return this.card.name + " added to discard";
-        case SHOW_ADD_DRAW:
-            return this.card.name + " added to draw";
-        case SHOW_ADD_HAND:
-            return this.card.name + " added to hand";
-        case SHOW_CARD_BRIEFLY:
-            return this.card.name + " shown briefly";
-        case SHOW_CARD_OBTAIN:
-        case FAST_OBTAIN:
-            return this.card.name + " added to deck";
-        default:
-            return "invalid card manipulation effect";
+            this.event = new TextEvent("Invalid card manipulation type; report to mod dev.");
         }
     }
 
     public void onAdd() {
-        Output.event(new TextEvent(this.getMessage()));
+        Output.event(this.event);
     }
 
     public static void registerHandlers(EffectManager manager) {